@@ -59,56 +68,67 @@
     "javascript-time-ago": "^0.4.8",
     "json-loader": "^0.5.4",
     "lodash": "^4.17.2",
+    "lodash-webpack-plugin": "^0.11.2",
     "offline-plugin": "^4.6.1",
     "postcss-cssnext": "^2.9.0",
     "postcss-inherit": "git+https://github.com/timsuchanek/postcss-inherit#build3",
     "postcss-inject": "^1.0.1",
     "postcss-simple-vars": "^3.0.0",
     "pre-push": "^0.1.1",
+    "purifycss-webpack": "^0.5.0",
     "rc-slider": "^5.4.3",
     "react": "^15.4.1",
     "react-algoliasearch": "^2.0.0",
-    "react-apollo": "^0.7.1",
+    "react-apollo": "1.0.0",
+    "react-async-bootstrapper": "^0.0.5",
+    "react-async-component": "git+https://github.com/timsuchanek/react-async-component#build4",
     "react-codemirror": "^0.3.0",
     "react-copy-to-clipboard": "^4.2.3",
     "react-dom": "^15.4.1",
-    "react-helmet": "^3.2.3",
+    "react-helmet": "5.0.2",
     "react-instantsearch": "^2.1.0",
+    "react-loadable": "^3.1.0",
     "react-modal": "^1.6.5",
     "react-redux": "^5.0.3",
+    "react-responsive": "^1.2.7",
     "react-router": "^2.8.0",
     "react-router-loader": "^0.5.4",
     "react-router-scroll": "^0.4.1",
+    "react-tree-walker": "2.0.0-alpha.3",
     "react-youtube": "^7.3.0",
     "redux": "^3.6.0",
+    "serialize-javascript": "^1.3.0",
     "slugify": "^1.1.0",
     "styled-components": "^1.1.3",
     "styled-jsx": "git+https://github.com/timsuchanek/styled-jsx#build",
-    "styled-jsx-postcss": "git+https://github.com/timsuchanek/styled-jsx-postcss#build",
+    "styled-jsx-postcss": "git+https://github.com/timsuchanek/styled-jsx-postcss#build3",
+    "svg-inline-loader": "^0.7.1",
     "typed-graphql": "^1.0.2",
     "webfontloader": "^1.6.27",
+    "webpack-bundle-analyzer": "^2.3.1",
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
     "@types/node": "^6.0.48",
     "awesome-typescript-loader": "git+https://github.com/timsuchanek/awesome-typescript-loader#build",
-    "css-loader": "^0.26.0",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.24.0",
+    "cheerio": "^0.22.0",
+    "css-loader": "0.28.0",
+    "file-loader": "0.10.1",
+    "html-webpack-plugin": "2.28.0",
     "image-webpack-loader": "^3.2.0",
     "lokka": "^1.7.0",
     "lokka-transport-http": "^1.6.1",
     "netlify-cli": "^1.1.0",
     "raw-loader": "^0.5.1",
     "react-hot-loader": "^3.0.0-beta.6",
-    "style-loader": "^0.13.1",
+    "style-loader": "0.16.1",
     "svgo-loader": "^1.1.2",
-    "tslint": "^4.0.0",
-    "tslint-config-standard": "^2.0.0",
+    "tslint": "4.0.0",
+    "tslint-config-standard": "2.0.0",
     "tslint-loader": "^3.3.0",
     "typescript": "^2.1.4",
     "url-loader": "^0.5.7",
-    "webpack": "^2.1.0-beta.27",
-    "webpack-dev-server": "^2.1.0-beta.11"
+    "webpack": "2.3.2",
+    "webpack-dev-server": "2.4.2"
   }
 }
