@@ -8,6 +8,7 @@ import styled from 'styled-components'
 import SimpleRelayTwin from './SimpleRelayTwin'
 import { breakpoints, maxWidth } from '../../../../utils/constants'
 const TimeAgo = require('javascript-time-ago').default
+import * as MediaQuery from 'react-responsive'
 
 require('javascript-time-ago/intl-messageformat-global')
 TimeAgo.locale(require('../../../../utils/locale'))
@@ -15,29 +16,6 @@ TimeAgo.locale(require('../../../../utils/locale'))
 interface Props {
   item: Item
 }
-
-const Root = styled.div`
-  position: relative;
-  display: flex;
-  max-width: 970px;
-  
-  @media (max-width: ${breakpoints.p580}px) {
-    margin-top: ${$v.size38};
-  }
-  
-  @media (max-width: 1050px) {
-    max-width: calc(100vw - 300px);
-  }
-  
-  @media (max-width: 750px) {
-    max-width: 750px;
-  }
-  
-`
-
-const RootContainer = styled.div`
-  max-width: ${maxWidth}px;
-`
 
 const Beta = styled.div`
   padding-top: 1px;
@@ -81,7 +59,6 @@ const Tag = styled.div `
 export default class ContentHeader extends React.Component<Props, {}> {
 
   render() {
-    const displayIcon = window.innerWidth > breakpoints.p900
     const {item} = this.props
     const lastModified = new Date(item.lastModified)
     const created = new Date(item.publicationDate)
@@ -109,18 +86,50 @@ export default class ContentHeader extends React.Component<Props, {}> {
 //    })()
 
     return (
-      <Root>
-        {displayIcon && (
+      <div className='root'>
+        <style jsx>{`
+          .root {
+            @p: .relative, .flex;
+            max-width: 970px;
+
+            @media (max-width: 580px) {
+              @p: .mt38;
+            }
+
+            @media (max-width: 1050px) {
+              max-width: calc(100vw - 300px);
+            }
+
+            @media (max-width: 750px) {
+              max-width: 750px;
+            }
+          }
+          .root-container {
+            @p: .pb60, .pt10, .w100, .bbox;
+            max-width: ${maxWidth}px;
+          }
+          @media (min-width: 1050px) {
+            .root-container.twin {
+              paddingRight: 232px;
+            }
+          }
+          .tag {
+            @p: .flex, .flexWrap;
+            @media (max-width: 580px) {
+              @p: .mt6;
+            }
+          }
+        `}</style>
+        <MediaQuery minWidth={900}>
           <IconWrapper className={cx($p.bbox, $p.db, $p.mr10, $p.relative)}>
             <CircleIcon width={44} height={44} type={item.layout}/>
           </IconWrapper>
-        )}
-        <RootContainer
-          className={cx($p.pb60, $p.pt10, $p.w100, $p.bbox)}
-          style={{
-            paddingRight: (layout === 'REFERENCE' && simpleRelayTwin && simpleRelayTwin.length > 0
-            && window.innerWidth > 1050) ? '232px' : '',
-          }}
+        </MediaQuery>
+        <div
+          className={cx(
+            'root-container',
+            {twin: layout === 'REFERENCE' && simpleRelayTwin && simpleRelayTwin.length > 0},
+          )}
         >
           <BreadCrumbContainer className={cx($p.ttu, $p.f14, $p.black20, $p.fw6)}>
             <BreadCrumb item={item} />
@@ -175,18 +184,14 @@ export default class ContentHeader extends React.Component<Props, {}> {
                 `Last updated ${timeAgo.format(lastModified)}`
               )}
             </div>
-            <div className={cx(
-              $p.flex,
-              $p.flexWrap,
-              window.innerWidth < breakpoints.p580 && $p.mt6,
-              )}>
-            {item.tags.map(tag => (
-              <Tag key={tag} className={cx($p.nowrap)}>#{tag}</Tag>
-            ))}
+            <div className='tag'>
+              {item.tags.map(tag => (
+                <Tag key={tag} className={cx($p.nowrap)}>#{tag}</Tag>
+              ))}
             </div>
           </Sublines>
-        </RootContainer>
-      </Root>
+        </div>
+      </div>
     )
   }
 }
