@@ -1,11 +1,10 @@
 import * as React from 'react'
-import { $p, $v } from 'graphcool-styles'
+import { $p } from 'graphcool-styles'
 import * as cx from 'classnames'
 import { Layout } from '../../../../types/types'
 import CircleIcon from '../../components/CircleIcon'
 import { Link } from 'react-router'
-import { breakpoints } from '../../../../utils/constants'
-import styled from 'styled-components'
+import * as MediaQuery from 'react-responsive'
 
 interface Props {
   layout: Layout
@@ -14,60 +13,65 @@ interface Props {
   link: string
 }
 
-const Root = styled.div`
-  padding: 0 ${$v.size60} 0 ${$v.size16};
-`
-
 const Heading = ({layout, title, text, link}: Props) => {
-  const displayIcon = window.innerWidth > breakpoints.p500
   return (
-    <Root>
+    <div>
+      <style jsx>{`
+        .heading {
+          padding: 0 60px 0 16px;
+        }
+        div :global(.link) {
+          @p: .flex, .itemsCenter, .noUnderline;
+          @media (min-width: 500px) {
+            @p: .ml4;
+          }
+        }
+        h1 {
+          @media (min-width: 500px) {
+            @p: .ml16
+          }
+        }
+        .text {
+          @media (max-width: 499px) {
+            @p: .mh10;
+          }
+          @media (min-width: 500px) {
+            @p: .ml60;
+          }
+        }
+      `}</style>
       {link.includes('http') ? (
         <a
           href={link}
-          className={cx(
-            $p.flex,
-            $p.itemsCenter,
-            $p.noUnderline,
-            displayIcon && $p.ml4,
-          )}>
-          {displayIcon && <CircleIcon
-            width={44}
-            height={44}
-            type={layout} />}
-          <h1
-            className={cx(
-              displayIcon && $p.ml16,
-            )}>{title}</h1>
+          className='link'
+        >
+          <MediaQuery minWidth={500}>
+            <CircleIcon
+              width={44}
+              height={44}
+              type={layout}
+            />
+          </MediaQuery>
+          <h1>{title}</h1>
         </a>
       ) : (
         <Link
           to={link}
-          className={cx(
-            $p.flex,
-            $p.itemsCenter,
-            $p.noUnderline,
-            displayIcon && $p.ml4,
-          )}>
-          {displayIcon && <CircleIcon
-            width={44}
-            height={44}
-            type={layout} />}
-          <h1
-            className={cx(
-              displayIcon && $p.ml16,
-              !displayIcon && $p.mh10,
-            )}>{title}</h1>
+          className='link'>
+          <MediaQuery minWidth={500}>
+            <CircleIcon
+              width={44}
+              height={44}
+              type={layout}
+            />
+          </MediaQuery>
+          <h1>{title}</h1>
         </Link>
       )}
-      <div
-        className={cx(
-          displayIcon && $p.ml60,
-          !displayIcon && $p.mh10,
-        )}>
+      <div className='text'>
         <p className={cx($p.mt25)}>{text}</p>
       </div>
-    </Root>
+    </div>
   )
 }
 
