@@ -1,38 +1,10 @@
-import * as React from 'react'
-import Header from '../Header/Header'
-import * as cx from 'classnames'
-import { $p } from 'graphcool-styles'
-import Footer from '../Footer/Footer'
-import FooterCTA from '../FooterCTA'
-import QueryEditor from './QueryEditor/QueryEditor'
-import * as Helmet from 'react-helmet'
+import { asyncComponent } from 'react-async-component'
 
-export default class FeaturesGraphQLView extends React.Component<{}, {}> {
-  componentDidMount() {
-    window.addEventListener('resize', this.rerender)
-  }
+const FeaturesGraphQLView = asyncComponent({
+  resolve: () => {
+    console.log('resolving features graphql view async')
+    return System.import('./FeaturesGraphQLViewAsync')
+  },
+})
 
-  render() {
-    return (
-      <div>
-        <Helmet title='GraphQL Database & API - Graphcool'/>
-        <Header view='HOMEPAGE' />
-        <QueryEditor ref='queryEditor' inViewPort={true}/>
-        <FooterCTA
-          headline1='Ready to get started?'
-          headline2='Set up a GraphQL backend or read the docs'
-          button1Text='Create GraphQL Backend'
-          button1Link='https://console.graph.cool/signup/'
-          button2Text='Open Docs'
-          button2Link='/docs/'
-          className={cx($p.mt96)}
-        />
-        <Footer/>
-      </div>
-    )
-  }
-
-  private rerender = () => {
-    this.forceUpdate()
-  }
-}
+export default FeaturesGraphQLView
